@@ -13,15 +13,17 @@
         "laravel"
     ],
     "require": {
-        "php": ">=7.2",
+        "php": ">=7.4",
         "ext-json": "*",
-        "illuminate/support": "5.7.*|5.8.*|^6.0|^7.0|^8.0",
-        "illuminate/http": "5.7.*|5.8.*|^6.0|^7.0|^8.0",
-        "illuminate/container": "5.7.*|5.8.*|^6.0|^7.0|^8.0",
-        "illuminate/pipeline": "5.7.*|5.8.*|^6.0|^7.0|^8.0"
+        "illuminate/support": "^6.0|^7.0|^8.0",
+        "illuminate/http": "^6.0|^7.0|^8.0",
+        "illuminate/container": "^6.0|^7.0|^8.0",
+        "illuminate/pipeline": "^6.0|^7.0|^8.0",
+        "doctrine/annotations": "^1.8",
+        "doctrine/cache": "^1.8"
     },
     "require-dev": {
-        "phpunit/phpunit": "^8.0",
+        "phpunit/phpunit": "^9.5",
         "mockery/mockery": "^1.0"
     },
     "autoload": {
