@@ -2,10 +2,16 @@
 
 namespace Tochka\JsonRpc;
 
+use Doctrine\Common\Annotations\AnnotationReader;
+use Doctrine\Common\Annotations\CachedReader;
+use Doctrine\Common\Cache\PhpFileCache;
+use Illuminate\Support\Facades\Config;
 use Illuminate\Support\ServiceProvider;
+use Tochka\JsonRpc\Casters\EnumCaster;
 use Tochka\JsonRpc\Exceptions\ExceptionHandler;
 use Tochka\JsonRpc\Support\JsonRpcHandleResolver;
 use Tochka\JsonRpc\Support\JsonRpcParser;
+use Tochka\JsonRpc\Support\JsonRpcRequestCast;
 
 class JsonRpcServiceProvider extends ServiceProvider
 {
@@ -16,17 +22,41 @@ class JsonRpcServiceProvider extends ServiceProvider
      */
     public function register(): void
     {
-        $this->app->singleton(\Tochka\JsonRpc\Facades\JsonRpcServer::class, static function () {
-            $parser = new JsonRpcParser();
-            $resolver = new JsonRpcHandleResolver();
+        $this->app->singleton(
+            Facades\JsonRpcRequestCast::class,
+            function () {
+                $reader = new CachedReader(
+                    new AnnotationReader(),
+                    new PhpFileCache($this->app->bootstrapPath('cache/annotations'), '.annotations.php'),
+                    Config::get('app.debug')
+                );
 
-            return new JsonRpcServer($parser, $resolver);
-        });
+                $instance = new JsonRpcRequestCast($reader);
+                if (class_exists('\BenSampo\Enum\Enum')) {
+                    $instance->addCaster(new EnumCaster());
+                }
+
+                return $instance;
+            }
+        );
+
+        $this->app->singleton(
+            \Tochka\JsonRpc\Facades\JsonRpcServer::class,
+            static function () {
+                $parser = new JsonRpcParser();
+                $resolver = new JsonRpcHandleResolver();
+
+                return new JsonRpcServer($parser, $resolver);
+            }
+        );
 
         // Обработчик ошибок JsonRpc
-        $this->app->singleton(\Tochka\JsonRpc\Facades\ExceptionHandler::class, static function () {
-            return new ExceptionHandler();
-        });
+        $this->app->singleton(
+            Facades\ExceptionHandler::class,
+            static function () {
+                return new ExceptionHandler();
+            }
+        );
     }
 
     /**
